requirements.txt
customtkinter
pillow
requests
geopy
python-dotenv
geocoder
